+++ Release_Notes.md
@@ -14,6 +14,7 @@
 ✅ Fixed https://github.com/byteben/Win32App-Migration-Tool/issues/18  
 ✅ Updated Licence terms to GNU GENERAL PUBLIC LICENSE  
 ✅ New module Get-LocalDetectionMethods to extract local detection methods (file/reg/msi) when detection is not a script  
+✅ New module New-IntuneDetectionMethod to create json for detection methods
 
 ## 2.0.19 - BETA - 16/12/2023