+++ Release_Notes.md
@@ -1,6 +1,6 @@
 # Win32App Migration Tool - Release Notes
 
-## 2.0.20 - BETA - 16/03/2024
+## 2.0.20 - BETA - 23/03/2024
 
 ✅ Export Detection Method to file in the working folder 'DetectionMethods' folder  
 ✅ Export Detection Method supporting information to Details\DeploymentTypes.csv  
@@ -15,6 +15,7 @@
 ✅ Updated Licence terms to GNU GENERAL PUBLIC LICENSE  
 ✅ New module Get-LocalDetectionMethods to extract local detection methods (file/reg/msi) when detection is not a script  
 ✅ New module New-IntuneDetectionMethod to create json for detection methods
+✅ Fixed an issue where icon export was attempted even if the icon was not present
 
 ## 2.0.19 - BETA - 16/12/2023